Fees are now using doubles instead of longs Upgrade feature written to upgrade existing dataset Added Booking by name query

+++ src/uk/org/woodcraft/bookings/pricing/PricingStrategy.java
@@ -4,5 +4,5 @@ import uk.org.woodcraft.bookings.datamodel.Booking;
 
 public interface PricingStrategy {
 
-	public long priceOf(Booking booking);
+	public double priceOf(Booking booking);
 }
